+++ src/main/java/mn/mosc/project/Application.java
@@ -1,0 +1,14 @@
+package mn.mosc.project;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * created by ubulgan on 9/30/17
+ */
+@SpringBootApplication
+public class Application {
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
