--- src/main/java/mn/mosc/project/web/ApplicationContextConfig.java
@@ -1,56 +1,0 @@
-package mn.mosc.project.web;
-
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
-import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.amazonaws.services.sns.AmazonSNS;
-import com.amazonaws.services.sns.AmazonSNSClientBuilder;
-import com.google.gson.Gson;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.ComponentScan.Filter;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-import org.springframework.context.annotation.FilterType;
-import org.springframework.stereotype.Controller;
-
-@Configuration
-@ComponentScan(basePackages = "mn.mosc.project",
-        excludeFilters = {
-                @Filter(value = Controller.class),
-                @Filter(type = FilterType.ASSIGNABLE_TYPE, value = ServletContextConfig.class)
-        })
-@SuppressWarnings("PMD.UseSingleton")
-@EnableAspectJAutoProxy
-public class ApplicationContextConfig {
-    //-----------Default max error retry is - 3 times:
-    @Bean
-    public AmazonS3 amazonS3() {
-        return AmazonS3ClientBuilder.defaultClient();
-    }
-
-    //-----------Default max error retry is - 3 times:
-    @Bean
-    public AmazonSNS amazonSNS() {
-        return AmazonSNSClientBuilder.defaultClient();
-    }
-
-    //-----------Default max error retry is - 10 times:
-    @Bean
-    public DynamoDBMapper dynamoDBMapper() {
-        return new DynamoDBMapper(AmazonDynamoDBClientBuilder.defaultClient());
-    }
-
-    //-----------Default max error retry is - 10 times:
-    @Bean
-    public AmazonDynamoDB amazonDynamoDB() {
-        return AmazonDynamoDBClientBuilder.defaultClient();
-    }
-
-    @Bean
-    public Gson gson() {
-        return new Gson();
-    }
-}
